[chore] 키보드w,s 기능 추가 및 vector int형으로 수정
변경
TitleScene
 - 키보드의 w,s를 눌렀을때도 반응하도록 수정

StageSelectScene
 - 키보드의 w,s를 눌렀을때도 반응하도록 수정

Util
 - 간단한 2차원 배열로 Vector2 를 int형으로 수정

+++ BabaIsYou/Util.cs
@@ -10,10 +10,10 @@ namespace BabaIsYou {
     }
 
     class Vector2 {
-        public double X { get; set; }
-        public double Y { get; set; }
+        public int X { get; set; }
+        public int Y { get; set; }
 
-        public Vector2(double x, double y) {
+        public Vector2(int x, int y) {
             X = x;
             Y = y;
         }
